requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
streamlit==1.29.0
requests==2.31.0
python-dotenv==1.0.0
plotly==5.18.0
seaborn==0.13.0
matplotlib==3.8.2
joblib==1.3.2
pytest==7.4.3
